+++ fabric/src/main/java/net/pl3x/map/fabric/client/Pl3xMapFabricClient.java
@@ -28,8 +28,6 @@ import java.util.concurrent.ExecutorService;
 import net.fabricmc.api.ClientModInitializer;
 import net.fabricmc.fabric.api.client.event.lifecycle.v1.ClientTickEvents;
 import net.fabricmc.fabric.api.client.keybinding.v1.KeyBindingHelper;
-import net.fabricmc.fabric.api.client.networking.v1.ClientConfigurationConnectionEvents;
-import net.fabricmc.fabric.api.client.networking.v1.ClientConfigurationNetworking;
 import net.fabricmc.fabric.api.client.networking.v1.ClientPlayConnectionEvents;
 import net.fabricmc.fabric.api.client.networking.v1.ClientPlayNetworking;
 import net.fabricmc.fabric.api.networking.v1.PayloadTypeRegistry;
@@ -42,10 +40,10 @@ import net.pl3x.map.core.network.Constants;
 import net.pl3x.map.core.scheduler.Scheduler;
 import net.pl3x.map.fabric.client.duck.MapInstance;
 import net.pl3x.map.fabric.client.manager.TileManager;
-import net.pl3x.map.fabric.main.network.ClientboundMapPayload;
-import net.pl3x.map.fabric.main.network.ClientboundServerPayload;
-import net.pl3x.map.fabric.main.network.ServerboundMapPayload;
-import net.pl3x.map.fabric.main.network.ServerboundServerPayload;
+import net.pl3x.map.fabric.common.network.ClientboundMapPayload;
+import net.pl3x.map.fabric.common.network.ClientboundServerPayload;
+import net.pl3x.map.fabric.common.network.ServerboundMapPayload;
+import net.pl3x.map.fabric.common.network.ServerboundServerPayload;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 import org.lwjgl.glfw.GLFW;
