+++ fabric/src/main/java/net/pl3x/map/fabric/common/network/ClientboundMapPayload.java
@@ -1,4 +1,4 @@
-package net.pl3x.map.fabric.main.network;
+package net.pl3x.map.fabric.common.network;
 
 import net.fabricmc.fabric.api.client.networking.v1.ClientPlayNetworking;
 import net.minecraft.client.Minecraft;
