+++ fabric/src/main/java/net/pl3x/map/fabric/common/network/ServerboundServerPayload.java
@@ -1,4 +1,4 @@
-package net.pl3x.map.fabric.main.network;
+package net.pl3x.map.fabric.common.network;
 
 import net.minecraft.network.FriendlyByteBuf;
 import net.minecraft.network.codec.StreamCodec;
